macro config

+++ src/kernel/arch/x86_64/include/panic.h
@@ -8,5 +8,5 @@ void _SystemRaiseHardError(const char* _file_, int line, const char* reason, con
 void PrintRegs(isr_state_t *reg);
 
 #define SystemRaiseHardError(reason, desc) _SystemRaiseHardError(__FILE__, __LINE__, reason, desc, NULL);
-#define ISR_SystemRaiseHardError(desc, regs) _SystemRaiseHardError(NULL, NULL, "ISR_STOP", desc, regs);
+#define ISR_SystemRaiseHardError(desc, regs) _SystemRaiseHardError(NULL, (int)NULL, "ISR_STOP", desc, regs);
 #define DebugRaiseKernelPanic() _SystemRaiseHardError(__FILE__, __LINE__, "DEBUG_STOP", "Manually invoked panic", NULL)
